//Mahesh

+++ src/main/java/com/monginis/ops/model/Orders.java
@@ -60,7 +60,7 @@ public class Orders {
 	
 	private int userId;
 	
-	private int isPositive;
+	private float isPositive;
 
 
 	private int menuId;
@@ -215,11 +215,13 @@ public class Orders {
 		this.userId = userId;
 	}
 
-	public int getIsPositive() {
+	
+
+	public float getIsPositive() {
 		return isPositive;
 	}
 
-	public void setIsPositive(int isPositive) {
+	public void setIsPositive(float isPositive) {
 		this.isPositive = isPositive;
 	}
